feat: use workspace folder as working directory by default
This allows better use of the docker shim, as otherwise shellcheck would
not be able to find the .shellcheckrc file on parent folders of the script.

This is potentially a breaking change, despite recent versions of shellcheck
handles source files a bit differently.

If you require the old behavior, set the configuration as false.