fix: prevent endless feedback loops
Add the `output` scheme to the default `shellcheck.ignoreFileSchemes`
list.

If `shellcheck.run` is set to `onType` (the default), writing to the log
may cause an endless feedback loop: VS Code treats the output channel as
a text document, and will emit a document-change event for every single
line logged to any output channel, which in turn may lead to our handler
logging even more lines, and so on.

This new check is arguably redundant: the language of the channel can
never be Shell, which means the existing `isAllowedTextDocument` check
happens to cover this case already, albeit by accident.

I still feel that there’s value in declaring this explicitly rather than
relying on accidental behavior.